Dev beto

+++ src/views/Pessoa/NovoEndereco.vue
@@ -217,6 +217,16 @@
         </div>
       </div>
     </form>
+    <b-modal
+      v-model="modalRemover"
+      title="Confirmar exclusão"
+      class="modal-danger"
+      ok-variant="danger"
+      @ok="ModalOk"
+      @hidden="ModalCancel"
+    >
+      Você confirma a exclusão desse registro?
+    </b-modal>
   </div>
 </template>
 
@@ -234,6 +244,8 @@ export default {
   },
   data() {
     return {
+      modalRemover: false,
+      itemRemover: null,
       tipos: [],
       loading: false,
       pagina: 1,
@@ -370,9 +382,17 @@ export default {
           });
         });
     },
-    Remover(id) {
+    ModalCancel(evento) {
+      evento.preventDefault();
+      this.itemRemover = null;
+    },
+    ModalOk(evento) {
+      evento.preventDefault();
+      this.modalRemover = false;
+      if (!this.itemRemover) return;
+
       this.$http({
-        url: "endereco/remover/" + id,
+        url: "endereco/remover/" + this.itemRemover,
         method: "DELETE"
       })
         .then(() => {
@@ -390,6 +410,10 @@ export default {
             duration: 10000
           });
         });
+    },
+    Remover(id) {
+      this.modalRemover = true;
+      this.itemRemover = id;
     },
     Novo() {
       this.loading = true;
